api/requirements.txt
Flask==3.0.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
python-dotenv==1.0.1
flask-cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
paho-mqtt==2.0.0